Made packet functions take const ptr, added format response packet function (#180)
* Made packet functions take const ptr, added format response packet function

* Remove packet.h, move everything into driver, use get_config for format_packet
