+++ cl_manycore/libraries/Makefile
@@ -9,7 +9,6 @@ HEADERS += bsg_manycore_loader.h
 HEADERS += bsg_manycore_tile.h
 HEADERS += bsg_manycore_mem.h
 HEADERS += bsg_manycore_mmio.h
-HEADERS += bsg_manycore_packet.h
 HEADERS += bsg_manycore_errno.h
 HEADERS += bsg_manycore_bits.h
 HEADERS += bsg_manycore_elf.h
